outreach_proj/requirements.txt
# Core Flask
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0

# OpenAI & Email
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
